feat(ui): remove examId from Content composable
This commit removes the `examId` parameter from the `Content` composable in the UI module
.
It also updates the `ImageContent` composable to directly use the text content of the item for the image path.
This change simplifies the `Content` composable and removes the dependency on the `examId`.

Additionally, the `examId` parameter is removed from the `Content` compos
able calls in the `CqScreen` and `CiScreen` composables in the `composequestion` and `composeinstruction` modules, respectively.
This further reduces the dependency on the `examId` and streamlines the code.