feat: add coverage report to PR
This commit adds a step to the build workflow to generate and upload coverage reports for JVM code.

The coverage report
 is generated using Kover and uploaded to the PR using the `mi-kas/kover-report` GitHub action.

This allows for easy visibility of code coverage changes and helps maintain a high level of test coverage.

Additionally, the commit uploads local coverage reports (XML and HTML) as artifacts for further
 analysis.

Finally, it removes duplicate coverage report generation and upload steps from the `androidTest` job.