feat(data): Add SubjectWithSeries data class
This commit introduces the `SubjectWithSeries` data class to the testing module. This class provides a convenient
 way to associate a `Subject` with its corresponding `Series`.

The `subjectWithSeries` list is created by mapping over the `subjects` list and creating a `SubjectWithSeries` instance for each subject, linking it to the appropriate series. This addition enhances the testing data and can be used for scenarios where
 both subject and series information are required.